+++ module-service/src/main/java/com/mjc/school/service/impl/NewsService.java
@@ -1,12 +1,15 @@
 package com.mjc.school.service.impl;
 
 import com.mjc.school.repository.BaseRepository;
+import com.mjc.school.repository.impl.AuthorModel;
 import com.mjc.school.repository.impl.NewsModel;
 import com.mjc.school.service.BaseService;
 import com.mjc.school.service.annotations.NotEmptyParam;
 import com.mjc.school.service.annotations.ValidParam;
 import com.mjc.school.service.dto.NewsDtoRequest;
 import com.mjc.school.service.dto.NewsDtoResponse;
+import com.mjc.school.service.exceptions.ErrorCode;
+import com.mjc.school.service.exceptions.ServiceException;
 import com.mjc.school.service.interfaces.NewsMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,10 +21,13 @@ import java.util.Optional;
 public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse, Long> {
 
     private final BaseRepository<NewsModel, Long> newsRepository;
+    private final BaseRepository<AuthorModel, Long> authorRepository;
 
     @Autowired
-    public NewsService(BaseRepository<NewsModel, Long> newsRepository) {
+    public NewsService(BaseRepository<NewsModel, Long> newsRepository,
+                       BaseRepository<AuthorModel, Long> authorRepository) {
         this.newsRepository = newsRepository;
+        this.authorRepository = authorRepository;
     }
 
     @Override
@@ -35,16 +41,20 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     @NotEmptyParam
     public NewsDtoResponse readById(Long id) {
         Optional<NewsModel> newsModel = newsRepository.readById(id);
-        NewsDtoResponse newsDtoResponse = new NewsDtoResponse();
-        if(newsModel.isPresent() && newsRepository.existById(id)) {
-            newsDtoResponse = NewsMapper.INSTANCE.newsToNewsDto(newsModel.get());
+        if(newsModel.isEmpty()) {
+            throw new ServiceException(String.format(
+                    ErrorCode.NOT_EXIST.getErrorMessage(), "News", id));
         }
-        return newsDtoResponse;
+        return NewsMapper.INSTANCE.newsToNewsDto(newsModel.get());
     }
 
     @Override
     @ValidParam
     public NewsDtoResponse create(NewsDtoRequest createRequest) {
+        if(!authorRepository.existById(createRequest.getAuthorId())) {
+            throw new ServiceException(String.format(
+                    ErrorCode.NOT_EXIST.getErrorMessage(), "Author", createRequest.getAuthorId()));
+        }
         NewsModel newsModel = newsRepository.create(NewsMapper.INSTANCE.newsDtoToNews(createRequest));
         return NewsMapper.INSTANCE.newsToNewsDto(newsModel);
     }
@@ -52,6 +62,10 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     @Override
     @ValidParam
     public NewsDtoResponse update(NewsDtoRequest updateRequest) {
+        if(!newsRepository.existById(updateRequest.getId())) {
+            throw new ServiceException(String.format(
+                    ErrorCode.NOT_EXIST.getErrorMessage(), "News", updateRequest.getId()));
+        }
         NewsModel newsModel = newsRepository.update(NewsMapper.INSTANCE.newsDtoToNews(updateRequest));
         return NewsMapper.INSTANCE.newsToNewsDto(newsModel);
     }
@@ -59,6 +73,9 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     @Override
     @NotEmptyParam
     public boolean deleteById(Long id) {
+        if(!newsRepository.existById(id)) {
+            throw new ServiceException(String.format(ErrorCode.NOT_EXIST.getErrorMessage(), "News", id));
+        }
         return newsRepository.deleteById(id);
     }
